backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asahi-backend"
version = "1.0.0"
description = "ASAHI LLM Inference Optimization Platform - Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1",
    "uvicorn[standard]>=0.32.0,<1",
    "pydantic>=2.0.0,<3",
    "pydantic-settings>=2.0.0,<3",
    "sqlalchemy[asyncio]>=2.0.0,<3",
    "asyncpg>=0.30.0,<1",
    "alembic>=1.14.0,<2",
    "redis[hiredis]>=5.0.0,<6",
    "stripe>=11.0.0,<12",
    "httpx>=0.27.0,<1",
    "PyJWT[crypto]>=2.8.0,<3",
    "celery[redis]>=5.4.0,<6",
    "resend>=2.0.0,<3",
    "openai>=1.0.0,<2",
    "anthropic>=0.39.0,<1",
    "sentence-transformers>=3.0.0,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9",
    "pytest-asyncio>=0.24.0,<1",
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
